Separate Controller from View
